requirements.txt
openai==0.27.4
PyYAML==6.0
requests==2.28.2
st_pages==0.3.5
streamlit==1.20.0
streamlit_lottie==0.0.3
